refactor: Improve code structure of the derive code
This also changes the implementation of `utils::field_aliases` to remove
the unnecessary `try_from`.
