+++ avro_derive/tests/expanded/avro_rs_207_rename_all_attribute.expanded.rs
@@ -29,7 +29,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                         name: "ITEM".to_string(),
                         doc: ::std::option::Option::None,
                         default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                        aliases: ::std::vec::Vec::new(),
+                        aliases: ::alloc::vec::Vec::new(),
                         schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                             named_schemas,
                             enclosing_namespace,
@@ -41,7 +41,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                         name: "DOUBLE_ITEM".to_string(),
                         doc: ::std::option::Option::None,
                         default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                        aliases: ::std::vec::Vec::new(),
+                        aliases: ::alloc::vec::Vec::new(),
                         schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                             named_schemas,
                             enclosing_namespace,
@@ -105,7 +105,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                 name: "ITEM".to_string(),
                 doc: ::std::option::Option::None,
                 default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                aliases: ::std::vec::Vec::new(),
+                aliases: ::alloc::vec::Vec::new(),
                 schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                     named_schemas,
                     enclosing_namespace,
@@ -117,7 +117,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                 name: "DOUBLE_ITEM".to_string(),
                 doc: ::std::option::Option::None,
                 default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                aliases: ::std::vec::Vec::new(),
+                aliases: ::alloc::vec::Vec::new(),
                 schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                     named_schemas,
                     enclosing_namespace,
