+++ avro_derive/tests/expanded/avro_rs_207_rename_attr_over_rename_all_attribute.expanded.rs
@@ -30,7 +30,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                         name: "ITEM".to_string(),
                         doc: ::std::option::Option::None,
                         default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                        aliases: ::std::vec::Vec::new(),
+                        aliases: ::alloc::vec::Vec::new(),
                         schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                             named_schemas,
                             enclosing_namespace,
@@ -42,7 +42,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                         name: "DoubleItem".to_string(),
                         doc: ::std::option::Option::None,
                         default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                        aliases: ::std::vec::Vec::new(),
+                        aliases: ::alloc::vec::Vec::new(),
                         schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                             named_schemas,
                             enclosing_namespace,
@@ -106,7 +106,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                 name: "ITEM".to_string(),
                 doc: ::std::option::Option::None,
                 default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                aliases: ::std::vec::Vec::new(),
+                aliases: ::alloc::vec::Vec::new(),
                 schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                     named_schemas,
                     enclosing_namespace,
@@ -118,7 +118,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                 name: "DoubleItem".to_string(),
                 doc: ::std::option::Option::None,
                 default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                aliases: ::std::vec::Vec::new(),
+                aliases: ::alloc::vec::Vec::new(),
                 schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                     named_schemas,
                     enclosing_namespace,
