+++ avro_derive/tests/expanded/avro_rs_501_namespace.expanded.rs
@@ -29,7 +29,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                         name: "a".to_string(),
                         doc: ::std::option::Option::None,
                         default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                        aliases: ::std::vec::Vec::new(),
+                        aliases: ::alloc::vec::Vec::new(),
                         schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                             named_schemas,
                             enclosing_namespace,
@@ -41,7 +41,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                         name: "b".to_string(),
                         doc: ::std::option::Option::None,
                         default: <String as ::apache_avro::AvroSchemaComponent>::field_default(),
-                        aliases: ::std::vec::Vec::new(),
+                        aliases: ::alloc::vec::Vec::new(),
                         schema: <String as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                             named_schemas,
                             enclosing_namespace,
@@ -106,7 +106,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                 name: "a".to_string(),
                 doc: ::std::option::Option::None,
                 default: <i32 as ::apache_avro::AvroSchemaComponent>::field_default(),
-                aliases: ::std::vec::Vec::new(),
+                aliases: ::alloc::vec::Vec::new(),
                 schema: <i32 as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                     named_schemas,
                     enclosing_namespace,
@@ -118,7 +118,7 @@ impl ::apache_avro::AvroSchemaComponent for A {
                 name: "b".to_string(),
                 doc: ::std::option::Option::None,
                 default: <String as ::apache_avro::AvroSchemaComponent>::field_default(),
-                aliases: ::std::vec::Vec::new(),
+                aliases: ::alloc::vec::Vec::new(),
                 schema: <String as ::apache_avro::AvroSchemaComponent>::get_schema_in_ctxt(
                     named_schemas,
                     enclosing_namespace,
